Driver deveolpment
Added init funciton
