+++ Heisprosjekt/driver/driver_test.py
@@ -1,4 +1,4 @@
-from elev import Elev
+from Elev import Elev
 
 
 elev = Elev()
